Implement m2m relations between Task and Label

+++ src/main/java/hexlet/code/service/LabelService.java
@@ -2,10 +2,12 @@ package hexlet.code.service;
 
 import hexlet.code.dto.label.LabelCreateDto;
 import hexlet.code.dto.label.LabelUpdateDto;
+import hexlet.code.exception.ResourceInUseDeleteException;
 import hexlet.code.exception.UniquenessViolationException;
 import hexlet.code.mapper.LabelMapper;
 import hexlet.code.model.Label;
 import hexlet.code.repository.LabelRepository;
+import hexlet.code.repository.TaskRepository;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
 
@@ -17,6 +19,8 @@ import java.util.Optional;
 public final class LabelService {
 
     private final LabelRepository repository;
+
+    private final TaskRepository taskRepository;
 
     private final LabelMapper mapper;
 
@@ -45,6 +49,9 @@ public final class LabelService {
     }
 
     public void delete(Long id) {
+        if (taskRepository.existsByLabelId(id)) {
+            throw new ResourceInUseDeleteException("Cannot delete. Label is referenced to one or more tasks.");
+        }
         repository.deleteById(id);
     }
